Refactor useIdle hook tests with comprehensive mock and test coverage
